Networkmessage methods

+++ src/tile.h
@@ -121,7 +121,7 @@ class TileItemVector : private ItemVector
 			}
 			return *getBeginDownItem();
 		}
-		void addDownItemCount(int32_t increment) {
+		void addDownItemCount(int16_t increment) {
 			downItemCount += increment;
 		}
